Add files via upload
Please we are adding this file concerning the call we had with Pelumi. We noticed that the folder we submitted didn't have our Jupyter notebook in the right extension so Pelumi instructed that we can submit it and make reference to the call we had with him as regards this. Thank you!